requirements.txt
opencv-python>=4.7.0
ultralytics>=8.2.0
torch>=1.12.0
numpy>=1.23.0
pandas>=1.4.0
facenet-pytorch>=2.5.3
deep-sort-realtime>=1.3.2
scipy>=1.9.0
matplotlib>=3.6.0
python-dotenv==1.0.0
streamlit==1.29.0
